+++ lib/module/isYesterday/isYesterday.d.ts
@@ -1,0 +1,12 @@
+interface isYesterdayType {
+    (date: string | Date | number): boolean;
+}
+/**
+ * 判断是否是昨天
+ * @param date
+ * @param options
+ * @returns {boolean}
+ * @version 2.6.0
+ */
+export declare const isYesterday: isYesterdayType;
+export {};