+++ lib/module/retry/retry.d.ts
@@ -1,0 +1,7 @@
+/**
+ * 重试
+ * @param cb
+ * @param times
+ * @param delay
+ */
+export declare function retry(cb: Function, times?: number, delay?: number): Promise<unknown>;