+++ lib/utils/now.d.ts
@@ -1,2 +1,0 @@
-declare const _now: number;
-export default _now;