+++ lib/utils/now.ts
@@ -1,3 +1,0 @@
-const _now = +new Date();
-
-export default _now;