+++ lib/utils/public.d.ts
@@ -1,6 +1,5 @@
 export declare const isNode: boolean;
 export declare const isBrowser: boolean;
-export declare const isHttps: boolean;
 /**
  * 内部默认空值检测
  * @param value